Put feedback on debrief page and wrote to separate file.
